Show truncated items, with option to expand them
A count is shown below files/dists where the full results are omitted,
searching for the specific thing will show it, which is aided by links
to that specific search.

Improves on issue #11, but still need to figure out a nicer way of
presenting the rather noticeably wrong count at the top right.