Calculate line numbers for the matched text
Note that the line numbers are not for the returned snippet of text, but
just the actual matching text.  This is more useful for accurate links
and the snippet's line boundaries may be calculated relative to the
matches line boundaries given the match's internal position within the
snippet.  Consider the "line" values to be the newline equivalent of the
"match" values.